Pr 45 cleanup (#46)
* Changed default http.Client && fix panic (#45)

* add case Image in mapToParams

* udpate gomod

* fix panic

* fix panic

* added an ability to changed client http

* add context in interface Downloader

* udpate gomod

* fix panic

* fix panic

* added an ability to changed client http

* add context in interface Downloader

* resolve conflict after rebase

* cleanup downloader integraion

---------
